+++ renderer/units/Unicode Foldings/import.js
@@ -3,12 +3,14 @@ const unit = document.getElementById ('unicode-foldings-unit');
 //
 const clearButton = unit.querySelector ('.clear-button');
 const charactersSamples = unit.querySelector ('.characters-samples');
+const countNumber = unit.querySelector ('.count-number');
 const loadButton = unit.querySelector ('.load-button');
 const saveButton = unit.querySelector ('.save-button');
 const charactersInput = unit.querySelector ('.characters-input');
 const codePointsInput = unit.querySelector ('.code-points-input');
 const useLocaleCheckbox = unit.querySelector ('.use-locale');
 const localeSelect = unit.querySelector ('.locale-select');
+const caseCountNumbers = unit.getElementsByClassName ('case-count-number');
 const charactersStrings = unit.getElementsByClassName ('characters-string');
 const codePointsStrings = unit.getElementsByClassName ('code-points-string');
 //
@@ -361,6 +363,7 @@ module.exports.start = function (context)
         event =>
         {
             let characters = event.currentTarget.value;
+            countNumber.textContent = Array.from (characters).length;
             codePointsInput.value = unicode.charactersToCodePoints (characters, true);
             for (let index = 0; index < charactersStrings.length; index++)
             {
@@ -368,6 +371,7 @@ module.exports.start = function (context)
                 let codePointsString = codePointsStrings[index];
                 charactersString.textContent = toCase (characters, charactersString.dataset.case, useLocaleCheckbox.checked ? localeSelect.value : undefined);
                 codePointsString.textContent = unicode.charactersToCodePoints (charactersString.textContent);
+                caseCountNumbers[index].textContent = Array.from (charactersString.textContent).length;
             }
         }
     );
@@ -380,6 +384,7 @@ module.exports.start = function (context)
         event =>
         {
             let characters = unicode.codePointsToCharacters (event.currentTarget.value);
+            countNumber.textContent = Array.from (characters).length;
             charactersInput.value = characters;
             for (let index = 0; index < charactersStrings.length; index++)
             {
@@ -387,6 +392,7 @@ module.exports.start = function (context)
                 let codePointsString = codePointsStrings[index];
                 charactersString.textContent = toCase (characters, charactersString.dataset.case, useLocaleCheckbox.checked ? localeSelect.value : undefined);
                 codePointsString.textContent = unicode.charactersToCodePoints (charactersString.textContent);
+                caseCountNumbers[index].textContent = Array.from (charactersString.textContent).length;
             }
         }
     );